hot fix

+++ 3-quick_sort.c
@@ -11,10 +11,10 @@
  *
  * Return: pivot final index
  */
-int partition(int *array, size_t low, size_t high, size_t size)
+int partition(int *array, int low, int high, size_t size)
 {
 	int pivot, temp;
-	size_t i, j;
+	int i, j;
 
 	pivot = array[high];
 	i = low;
